arxiv-api/requirements.txt
lxml==4.1.0
Cython==0.28.2
numpy==1.14.2 
matplotlib==1.5
scipy==0.17
psycopg2
POT==0.4.0
tensorflow==1.7.0
Django==2.0.4
nltk==3.2.4
fastText==0.8.3
djangorestframework==3.8.2
scikit_learn==0.19.1
django-cors-headers